CSV 파싱 및 UI 동기화 로직 구현 완료

+++ Assets/Scripts/DataType/StatData.cs
@@ -11,7 +11,7 @@ public class StatData : ScriptableObject
     [TextArea] 
     public string Description;
     public StatCategory StatCategory;
-    public float MinValue;
-    public float MaxValue;
+    public int MinValue;
+    public int MaxValue;
     public string IconPath;
 }
